Supported cultures

+++ src/services/CCRS.Catalog.API/Models/Feedback.cs
@@ -8,7 +8,7 @@ namespace CCRS.Catalog.API.Models
     {
         public int Id { get; set; }
         public int UserId { get; set; }        
-        public int Stars { get; set; }
+        public float Stars { get; set; }
         public string Comment { get; set; }
         public string Image { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.Now;
